refactor: pluralize package names
