+++ src/main/java/org/sosly/vlp/events/capabilities/VillagerStatusEventHandler.java
@@ -1,4 +1,4 @@
-package org.sosly.vlp.capability;
+package org.sosly.vlp.events.capabilities;
 
 import net.minecraftforge.common.capabilities.RegisterCapabilitiesEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
@@ -7,7 +7,7 @@ import org.sosly.vlp.VillagerLiteracyProgram;
 import org.sosly.vlp.api.capability.IVillagerStatus;
 
 @Mod.EventBusSubscriber(modid = VillagerLiteracyProgram.MOD_ID, bus = Mod.EventBusSubscriber.Bus.MOD)
-public class CapabilityRegistry {
+public class VillagerStatusEventHandler {
     @SubscribeEvent
     public static void registerCapabilities(RegisterCapabilitiesEvent event) {
         event.register(IVillagerStatus.class);
